+++ R/connect.R
@@ -92,13 +92,18 @@ Client <- R6::R6Class( # nolint
             "; the server did not provide its version."
           )
         )
-      } else if (compareVersion(settings$version, self$minimum_server_version) < 0) {
-        stop(
-          glue::glue("ERROR: Requires Posit Connect server version >= ",
-            self$minimum_server_version,
-            ", current version ",
-            settings$version)
-        )
+      } else {
+        # Drop the -dev version suffix, if present
+        parsed_version <- sub("-.*$", "", settings$version)
+
+        if (compareVersion(parsed_version, self$minimum_server_version) < 0) {
+          stop(
+            glue::glue("ERROR: Requires Posit Connect server version >= ",
+              self$minimum_server_version,
+              ", current version ",
+              settings$version)
+          )
+        }
       }
     },
     raise_error = function(res) {